cassandra consul scripts
