+++ packs/bitesize/actions/cass_consul.yaml
@@ -1,0 +1,10 @@
+name: cass_consul
+pack: "bitesize"
+runner_type: run-python
+description: "creates a consul acl rule"
+enabled: true
+entry_point: "cass_consul.py"
+parameters:
+  cluster:
+    type: string
+    required: true
